// src/shared/myStyles/myStyles.d.ts
export declare const textColors: {
    primary: string;
    main: string;
    secondary: string;
    link: string;
    formLabel: string;
    highlight: string;
    placeholder: string;
};
export declare const buttonStyles: {
    main: string;
    paginationActive: string;
    paginationPassive: string;
    navActive: string;
    navPassive: string;
    close: string;
};
export declare const borderColors: {
    primary: string;
    formInput: string;
};
export declare const pageTitleStyle: string;